+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 5.4.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Customize INameChooser adapter to check also alias ids and disallow to create contents that could override aliases.
+  [cekk]
+- Add flag in controlpanel to enable/disable INameChooser customization.
+  [cekk]
 
 
 5.4.8 (2024-03-19)